requirements.txt
torch
torchvision
numpy
matplotlib
pillow
scikit-image
tqdm
scipy
gradio
